last bit of refactoring for now: hint types
Move all types actually intended to be used for type-hinting into
a new module 'hint_types'.  Still want to refactor:
- structured/basic_types.py
- structured/utils.py
As they share things in common now, but needs a little thinking to avoid
circular imports.
