+++ structured/__init__.py
@@ -3,6 +3,6 @@ __version__ = '3.0.0'
 
 from .base_types import *
 from .basic_types import *
-from .complex_types import *
+from .hint_types import *
 from .serializers import *
 from .structured import *
