--- structured/complex_types/__init__.py
@@ -1,2 +1,0 @@
-from .arrays import Header, array
-from .strings import NET, EncoderDecoder, char, null_char, null_unicode, unicode
